+++ code/HapticGlove/lib/GloveControl/GloveControl.cpp
@@ -1,12 +1,13 @@
 #include "GloveControl.h"
 
 void gloveControlSetup(){
+  // initialize IMUs, Hall-Effect Sensors, etc
     return;
 }
 
 void sendPositionData(){
   uint8_t fpos[16];
-  uint8_t wpos[3];
+  uint32_t wpos[3];
   uint8_t apos[3];
   for(int j=0; j<16; j++){
     fpos[j] = random(1, 255);
